+++ netty/src/main/java/io/grpc/netty/InternalHandlerSettings.java
@@ -17,60 +17,28 @@
 package io.grpc.netty;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Preconditions;
 import io.grpc.Internal;
 
 /**
- * Allows autoFlowControl to be turned on and off from interop testing and flow control windows to
- * be accessed. For internal use only.
+ * Controlled accessor to {@link NettyHandlerSettings}.
  */
 @VisibleForTesting // Visible for tests in other packages.
 @Internal
 public final class InternalHandlerSettings {
 
-  private static volatile boolean enabled;
-
-  private static boolean autoFlowControlOn;
-  // These will be the most recently created handlers created using NettyClientTransport and
-  // NettyServerTransport
-  private static AbstractNettyHandler clientHandler;
-  private static AbstractNettyHandler serverHandler;
-
-  static void setAutoWindow(AbstractNettyHandler handler) {
-    if (!enabled) {
-      return;
-    }
-    synchronized (InternalHandlerSettings.class) {
-      handler.setAutoTuneFlowControl(autoFlowControlOn);
-      if (handler instanceof NettyClientHandler) {
-        clientHandler = handler;
-      } else if (handler instanceof NettyServerHandler) {
-        serverHandler = handler;
-      } else {
-        throw new RuntimeException("Expecting NettyClientHandler or NettyServerHandler");
-      }
-    }
-  }
-
   public static void enable(boolean enable) {
-    enabled = enable;
+    NettyHandlerSettings.enable(enable);
   }
 
   public static synchronized void autoWindowOn(boolean autoFlowControl) {
-    autoFlowControlOn = autoFlowControl;
+    NettyHandlerSettings.autoWindowOn(autoFlowControl);
   }
 
   public static synchronized int getLatestClientWindow() {
-    return getLatestWindow(clientHandler);
+    return NettyHandlerSettings.getLatestClientWindow();
   }
 
   public static synchronized int getLatestServerWindow() {
-    return getLatestWindow(serverHandler);
-  }
-
-  private static synchronized int getLatestWindow(AbstractNettyHandler handler) {
-    Preconditions.checkNotNull(handler);
-    return handler.decoder().flowController()
-        .initialWindowSize(handler.connection().connectionStream());
+    return NettyHandlerSettings.getLatestServerWindow();
   }
 }
