+++ netty/src/main/java/io/grpc/netty/NettyServerTransport.java
@@ -85,7 +85,7 @@ class NettyServerTransport implements ServerTransport {
 
     // Create the Netty handler for the pipeline.
     final NettyServerHandler grpcHandler = createHandler(listener);
-    InternalHandlerSettings.setAutoWindow(grpcHandler);
+    NettyHandlerSettings.setAutoWindow(grpcHandler);
 
     // Notify when the channel closes.
     channel.closeFuture().addListener(new ChannelFutureListener() {
